algunos test

+++ src/main/java/com/chriscorp/crud/controller/UserController.java
@@ -1,12 +1,13 @@
 package com.chriscorp.crud.controller;
 
+import com.chriscorp.crud.exception.ModelNotFoundException;
 import com.chriscorp.crud.model.User;
 import com.chriscorp.crud.service.userImpl.UserServiceImpl;
-import lombok.Data;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
-import java.nio.file.OpenOption;
 import java.util.List;
 import java.util.Optional;
 
@@ -18,28 +19,39 @@ public class UserController {
     private UserServiceImpl userService;
 
     @GetMapping
-    public List<User> getAllUser(){
-        return userService.getAllUser();
+    public ResponseEntity<List<User>> getAllUser(){
+        return new ResponseEntity<>(userService.getAllUser(),HttpStatus.OK);
     }
 
     @GetMapping("/{id}")
-    public Optional<User> getUserById(@PathVariable(name = "id") long id){
-        return userService.getUserById(id);
+    public ResponseEntity<Optional<User>> getUserById(@PathVariable(name = "id") long id) throws Exception{
+        Optional<User> user = userService.getUserById(id);
+        if(user.isPresent()) {
+            return new ResponseEntity<>(userService.getUserById(id), HttpStatus.OK);
+        } else {
+            throw new ModelNotFoundException(id);
+        }
     }
 
     @DeleteMapping("/{id}")
-    public void deleteUserById(@PathVariable(name = "id")  long id){
-        userService.deleteUserById(id);
+    public ResponseEntity<Object> deleteUserById(@PathVariable(name = "id")  long id) throws Exception{
+        Optional<User> user = userService.getUserById(id);
+        if(user.isPresent()) {
+            userService.deleteUserById(id);
+            return new ResponseEntity<>(HttpStatus.OK);
+        } else {
+            throw new ModelNotFoundException(id);
+        }
     }
 
     @PostMapping
-    public User postUser(@RequestBody User user){
-        return userService.postUser(user);
+    public ResponseEntity<User> postUser(@RequestBody User user){
+        return new ResponseEntity<>(userService.postUser(user),HttpStatus.OK) ;
     }
 
     @PutMapping("/{id}")
-    public User putUser(@RequestBody User user, @PathVariable(name = "id")  long id) throws Exception {
-        return userService.getUserById(id).map(newObj -> {
+    public ResponseEntity<User> putUser(@RequestBody User user, @PathVariable(name = "id")  long id) throws Exception {
+        return new ResponseEntity<>(userService.getUserById(id).map(newObj -> {
             newObj.setNombre(user.getNombre());
             newObj.setApellidos(user.getApellidos());
             newObj.setEdad(user.getEdad());
@@ -47,7 +59,8 @@ public class UserController {
             newObj.setEmail(user.getEmail());
             newObj.setTelefono(user.getTelefono());
             return userService.postUser(newObj);
-        }).orElseThrow(() -> new Exception("Product with id " + id + "not found"));
+        }).orElseThrow(() -> new Exception("Product with id " + id + "not found")),
+        HttpStatus.OK);
     }
 
 }
